Implement gallery snippet
The gallery is built with @lokesh/lightbox2, @zroger/jekyll-minimagick, and
@jgatjens loop_directory plugin. https://gist.github.com/jgatjens/8925165

Lightbox was modified to make the icons retina-ready and get rid of the
animations. MiniMagick was not modified. loop_directory was modified to look
up the path as a variable instead of a filepath if the file does not exist
on disk (working around Liquid's stupidity).